refactor: organize DTOs into domain-specific subfolders
