+++ src/Dto/Auth/SessionDto.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Rjds\PhpLastfmClient\Dto;
+namespace Rjds\PhpLastfmClient\Dto\Auth;
 
 use Rjds\PhpDto\Attribute\CastTo;
