+++ src/Dto/Common/ImageDto.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Rjds\PhpLastfmClient\Dto;
+namespace Rjds\PhpLastfmClient\Dto\Common;
 
 use Rjds\PhpDto\Attribute\MapFrom;
