+++ src/Dto/Library/LibraryArtistDto.php
@@ -2,11 +2,12 @@
 
 declare(strict_types=1);
 
-namespace Rjds\PhpLastfmClient\Dto;
+namespace Rjds\PhpLastfmClient\Dto\Library;
 
 use Rjds\PhpDto\Attribute\ArrayOf;
 use Rjds\PhpDto\Attribute\CastTo;
 use Rjds\PhpDto\Attribute\MapFrom;
+use Rjds\PhpLastfmClient\Dto\Common\ImageDto;
 
 final readonly class LibraryArtistDto
 {
