+++ src/Dto/User/UserDto.php
@@ -2,11 +2,12 @@
 
 declare(strict_types=1);
 
-namespace Rjds\PhpLastfmClient\Dto;
+namespace Rjds\PhpLastfmClient\Dto\User;
 
 use Rjds\PhpDto\Attribute\ArrayOf;
 use Rjds\PhpDto\Attribute\CastTo;
 use Rjds\PhpDto\Attribute\MapFrom;
+use Rjds\PhpLastfmClient\Dto\Common\ImageDto;
 
 final readonly class UserDto
 {
